Update VueTagInput.vue
Add deleteOnBackspace prop to allow turning off "delete tag if key backspace at the start of input"

+++ src/VueTagInput.vue
@@ -120,6 +120,11 @@ export default {
       required: false,
       default: false,
     },
+    deleteOnBackspace: {
+      type: Boolean,
+      required: false,
+      default: true,
+    },
   },
   data() {
     return {
@@ -199,7 +204,7 @@ export default {
     // Event sequence: keydown -> compositionstart -> input -> compositionend -> keyup
     handleKeydown(e) {
       // delete tag if key backspace at the start of input
-      if (e.keyCode === KEYS.BACKSPACE && this.$refs.input.selectionStart === 0) {
+      if (this.deleteOnBackspace && e.keyCode === KEYS.BACKSPACE && this.$refs.input.selectionStart === 0) {
         this.deleteTag(this.tags.length - 1);
       }
